+++ src/Exception/Api/SlotException.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * Copyright (c) velkuns
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Velkuns\ArtifactsMMO\Exception\Api;
+
+use Velkuns\ArtifactsMMO\Exception\ArtifactsMMOApiException;
+
+/**
+ * Code error: 491
+ */
+class SlotException extends ArtifactsMMOApiException {}