+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ----
 
+## [1.1.0] - 2024-08-22
+[1.1.0]: https://github.com/velkuns/artifactsmmo-client/compare/1.0.2...1.1.0
+### Added
+- Add new Custom Api Exception based on error from API
+- PHP 8.3 support
+### Changed
+- Now handle retry request when timeout or when bank / G.E already are in transaction
+
 ## [1.0.2] - 2024-08-11
 [1.0.2]: https://github.com/velkuns/artifactsmmo-client/compare/1.0.1...1.0.2
 ### Changed